requirements.txt
azure-ai-inference
azure-search-documents
tiktoken
requests
python-dotenv
openai
